`GLfloat` -> `float`.
Cleaning up code.

	modified:   code/ajokki/ajokki.cpp
	modified:   code/hirvi/hirvi.cpp
	modified:   code/ylikuutio/geometry/transformation.cpp
	modified:   code/ylikuutio/ontology/camera.cpp
	modified:   code/ylikuutio/ontology/model.hpp
	modified:   code/ylikuutio/ontology/text3D.hpp
	modified:   code/ylikuutio/ontology/universe.cpp
	modified:   code/ylikuutio/ontology/universe.hpp
	modified:   code/ylikuutio/triangulation/triangulation_templates.hpp

+++ code/ylikuutio/geometry/transformation.cpp
@@ -41,9 +41,9 @@ namespace yli
             spherical_vertex.phi = DEGREES_TO_RADIANS(-planar_world_vertex.z);  // phi is latitude, the polar angle.
 
             glm::vec3 cartesian_vertex;
-            cartesian_vertex.x = (GLfloat) (spherical_vertex.rho * sin(spherical_vertex.theta) * cos(spherical_vertex.phi));
-            cartesian_vertex.y = (GLfloat) (spherical_vertex.rho * sin(spherical_vertex.theta) * sin(spherical_vertex.phi));
-            cartesian_vertex.z = (GLfloat) (spherical_vertex.rho * cos(spherical_vertex.theta));
+            cartesian_vertex.x = spherical_vertex.rho * sin(spherical_vertex.theta) * cos(spherical_vertex.phi);
+            cartesian_vertex.y = spherical_vertex.rho * sin(spherical_vertex.theta) * sin(spherical_vertex.phi);
+            cartesian_vertex.z = spherical_vertex.rho * cos(spherical_vertex.theta);
 
             return cartesian_vertex;
         }
@@ -92,8 +92,8 @@ namespace yli
                 for (uint32_t x = 0; x < image_width; x++)
                 {
                     glm::vec3 spherical_terrain_vertex = temp_vertices[temp_vertices_i];
-                    spherical_terrain_vertex.x = (GLfloat) current_longitude_in_degrees;
-                    spherical_terrain_vertex.z = (GLfloat) current_latitude_in_degrees;
+                    spherical_terrain_vertex.x = static_cast<float>(current_longitude_in_degrees);
+                    spherical_terrain_vertex.z = static_cast<float>(current_latitude_in_degrees);
                     temp_vertices[temp_vertices_i++] = yli::geometry::transform_planar_world_vertex_into_cartesian_vertex(spherical_terrain_vertex, sphere_radius);
 
                     current_longitude_in_degrees += longitude_step_in_degrees;
@@ -126,8 +126,8 @@ namespace yli
                     for (std::size_t x = 1; x < image_width; x++)
                     {
                         glm::vec3 spherical_terrain_vertex = temp_vertices[temp_vertices_i];
-                        spherical_terrain_vertex.x = (GLfloat) current_longitude_in_degrees;
-                        spherical_terrain_vertex.z = (GLfloat) current_latitude_in_degrees;
+                        spherical_terrain_vertex.x = static_cast<float>(current_longitude_in_degrees);
+                        spherical_terrain_vertex.z = static_cast<float>(current_latitude_in_degrees);
                         temp_vertices[temp_vertices_i++] = yli::geometry::transform_planar_world_vertex_into_cartesian_vertex(spherical_terrain_vertex, sphere_radius);
 
                         current_longitude_in_degrees += longitude_step_in_degrees;
